Update gravatar.html
The current version of gravatar-image will only use the username or hash defined in the element's prototype if a gravatar-image element is being modified after creation (i.e. in a repeating template).These changes will allow you to modify the size, username, and hash inside a repeating template and will yield the results that you are looking for in this context.